+++ src/main/resources/application-test.properties
@@ -12,5 +12,5 @@ spring.security.oauth2.client.registration.google.client-secret=mock_google_clie
 spring.security.oauth2.client.registration.google.redirect-uri=http://localhost:8080/login/oauth2/code/google
 spring.security.oauth2.client.registration.google.scope=https://www.googleapis.com/auth/userinfo.profile
 spring.security.oauth2.client.provider.google.authorization-uri=https://accounts.google.com/o/oauth2/v2/auth
-spring.security.oauth2.client.provider.google.token-uri=http://localhost:8089/o/oauth2/token
+spring.security.oauth2.client.provider.google.token-uri=http://localhost:8089/token
 spring.security.oauth2.client.provider.google.user-info-uri=http://localhost:8089/oauth2/v2/userinfo
