+++ src/main/resources/application.properties
@@ -1,18 +1,18 @@
 spring.security.oauth2.client.registration.github.provider=github
-spring.security.oauth2.client.registration.github.client-id=Ov23liTBhugSIcf8VX1v
+spring.security.oauth2.client.registration.github.client-id=Ov23liIHA0fth835Cwdc
 #TODO: client_secret
-spring.security.oauth2.client.registration.github.client-secret=your_client_secret
+spring.security.oauth2.client.registration.github.client-secret=my_secret
 spring.security.oauth2.client.registration.github.redirect-uri=http://localhost:8080/login/oauth2/code/github
 spring.security.oauth2.client.registration.github.scope=read:user
 spring.security.oauth2.client.provider.github.authorization-uri=https://github.com/login/oauth/authorize
 spring.security.oauth2.client.provider.github.token-uri=https://github.com/login/oauth/access_token
 spring.security.oauth2.client.provider.github.user-info-uri=https://api.github.com/user
 spring.security.oauth2.client.registration.google.provider=google
-spring.security.oauth2.client.registration.google.client-id=88951150265-q7aq5urrim4rrtqo62hc96o6r5v7r34c.apps.googleusercontent.com
+spring.security.oauth2.client.registration.google.client-id=441784291596-au4td16867s4551nfmvgc8hfm3ftl75a.apps.googleusercontent.com
 #TODO: client_secret
-spring.security.oauth2.client.registration.google.client-secret=your_client_secret
+spring.security.oauth2.client.registration.google.client-secret=my_secret
 spring.security.oauth2.client.registration.google.redirect-uri=http://localhost:8080/login/oauth2/code/google
-spring.security.oauth2.client.registration.google.scope=https://www.googleapis.com/auth/userinfo.profile https://www.googleapis.com/auth/userinfo.email
+spring.security.oauth2.client.registration.google.scope=https://www.googleapis.com/auth/userinfo.profile
 spring.security.oauth2.client.provider.google.authorization-uri=https://accounts.google.com/o/oauth2/auth
-spring.security.oauth2.client.provider.google.token-uri=https://accounts.google.com/o/oauth2/token
-spring.security.oauth2.client.provider.google.user-info-uri=https://www.googleapis.com/oauth2/v1/userinfo
+spring.security.oauth2.client.provider.google.token-uri=https://accounts.google.com/token
+spring.security.oauth2.client.provider.google.user-info-uri=https://www.googleapis.com/oauth2/v2/userinfo
